requirements.txt
# AI Benchmark Suite Requirements

# Core dependencies
pyyaml>=6.0
requests>=2.28.0
numpy>=1.21.0
pandas>=1.3.0

# Optional API dependencies
openai>=1.0.0
anthropic>=0.3.0

# Logging and utilities
tqdm>=4.64.0
click>=8.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0